requirements.txt
bokeh==1.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.5
packaging==20.8
pandas
Pillow==8.1.0
pip==20.3.3
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.25.1
setuptools==41.2.0
simplejson==3.17.2
six==1.15.0
tornado==6.1
typing-extensions==3.7.4.3
urllib3==1.26.2
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3